backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-dotenv==1.0.1
groq==0.9.0
gtts==2.5.1
moviepy==1.0.3
Pillow==10.3.0
requests==2.31.0
pydantic==2.7.1
python-multipart==0.0.9
aiofiles==23.2.1
httpx==0.27.0
numpy==1.26.4
edge-tts==6.1.12
imageio==2.34.1
imageio-ffmpeg==0.5.1
openai==1.35.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
cloudinary==1.40.0
hypothesis==6.103.2
pytest==8.2.2
pytest-asyncio==0.23.7
